Added picom config, some new aliases, removed tmux
There was an issue with videos/scrolling on chromium as well as when
playing videos, the display was choppy. Took me a while to figure out
a fix for that, and I think I've fixed it now.

Also I've added some new aliases, they are most likely temporary.

Also also, before this, tmux was launched right after I launched the
terminal, and most of the time, I would end up not using it at all. So,
removed that, and now I'll just use it when I need it.

Also, I again modified that `solarized-modified` colorscheme on
konsole,but currently I don't have a way to convert it to kitty.
So, adding it too.